requirements.txt
absl-py==1.4.0
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==2.0.12
colorama==0.4.5
cycler==0.11.0
dataclasses==0.8
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.48.2
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
imutils==0.5.4
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
numpy==1.19.5
oauthlib==3.2.2
opencv-python==4.6.0.66
pandas==1.1.5
Pillow==8.4.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytesseract==0.3.8
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.5.4
seaborn==0.11.2
six==1.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
torch==1.9.1
torchvision==0.10.1
tqdm==4.64.1
typing_extensions==4.1.1
urllib3==1.26.14
Werkzeug==2.0.3
zipp==3.6.0
